Add support for sparse matrices (#10)
* support scipy.sparse matrices
* separate utility functions into `utilities.py`
