+++ chord_diagram.py
@@ -10,23 +10,213 @@ from matplotlib.colors import ColorConverter
 from matplotlib.path import Path
 
 import numpy as np
+import scipy.sparse as ssp
 
 from .gradient import gradient
+from .utilities import _get_normed_line, dist, polar2xy
 
 
 LW = 0.3
 
 
-def dist(points):
-    x1, y1 = points[0]
-    x2, y2 = points[1]
+def chord_diagram(mat, names=None, width=0.1, pad=2., gap=0., chordwidth=0.7,
+                  ax=None, colors=None, cmap=None, alpha=0.7,
+                  use_gradient=False, show=False, **kwargs):
+    """
+    Plot a chord diagram.
 
-    return np.sqrt((x2 - x1)*(x2 - x1) + (y2 - y1)*(y2 - y1))
+    Parameters
+    ----------
+    mat : square matrix
+        Flux data, mat[i, j] is the flux from i to j
+    names : list of str, optional (default: no names)
+        Names of the nodes that will be displayed.
+    width : float, optional (default: 0.1)
+        Width/thickness of the ideogram arc.
+    pad : float, optional (default: 2)
+        Distance between two neighboring ideogram arcs. Unit: degree.
+    gap : float, optional (default: 0)
+        Distance between the arc and the beginning of the cord.
+    chordwidth : float, optional (default: 0.7)
+        Position of the control points for the chords, controlling their shape.
+    ax : matplotlib axis, optional (default: new axis)
+        Matplotlib axis where the plot should be drawn.
+    colors : list, optional (default: from `cmap`)
+        List of user defined colors or floats.
+    cmap : str or colormap object (default: viridis)
+        Colormap to use.
+    alpha : float in [0, 1], optional (default: 0.7)
+        Opacity of the chord diagram.
+    use_gradient : bool, optional (default: False)
+        Whether a gradient should be use so that chord extremities have the
+        same color as the arc they belong to.
+    **kwargs : keyword arguments
+        Available kwargs are "fontsize" and "sort" (either "size" or
+        "distance"), "zero_entry_size" (in degrees, default: 0.5).
+    """
+    import matplotlib.pyplot as plt
+
+    if ax is None:
+        _, ax = plt.subplots()
+
+    # copy matrix and set a minimal value for visibility of zero fluxes
+    is_sparse = ssp.issparse(mat)
+
+    if is_sparse:
+        mat = mat.tocsr(copy=True)
+    else:
+        mat = np.array(mat, copy=True)
+
+    # mat[i, j]:  i -> j
+    num_nodes = mat.shape[0]
+
+    # set entries for zero entries that have a nonzero reciprocal
+    min_deg  = kwargs.get("zero_entry_size", 0.5)
+    min_deg *= mat.sum() / (360 - num_nodes*pad)
+
+    if is_sparse:
+        nnz = mat.nonzero()
+
+        for i, j in zip(*nnz):
+            if mat[j, i] == 0:
+                mat[j, i] = min_deg
+    else:
+        zeros = np.argwhere(mat == 0)
+
+        for (i, j) in zeros:
+            if mat[j, i] != 0:
+                mat[i, j] = min_deg
+
+    # sum over rows
+    x = mat.sum(axis=1).A1 if is_sparse else mat.sum(axis=1)
+
+    print(x)
+
+    # configure colors
+    if colors is None:
+        colors = np.linspace(0, 1, num_nodes)
+
+    if cmap is None:
+        cmap = "viridis"
+
+    if isinstance(colors, (Sequence, np.ndarray)):
+        assert len(colors) == num_nodes, "One color per node is required."
+
+        # check color type
+        first_color = colors[0]
+
+        if isinstance(first_color, (int, float, np.integer)):
+            cm = plt.get_cmap(cmap)
+            colors = cm(colors)[:, :3]
+        else:
+            colors = [ColorConverter.to_rgb(c) for c in colors]
+    else:
+        raise ValueError("`colors` should be a list.")
+
+    # find position for each start and end
+    y = x / np.sum(x).astype(float) * (360 - pad*len(x))
+
+    pos = {}
+    arc = []
+    nodePos = []
+    start = 0
+
+    for i in range(num_nodes):
+        end = start + y[i]
+        arc.append((start, end))
+        angle = 0.5*(start+end)
+
+        if -30 <= angle <= 210:
+            angle -= 90
+        else:
+            angle -= 270
+
+        nodePos.append(
+            tuple(polar2xy(1.1, 0.5*(start + end)*np.pi/180.)) + (angle,))
+
+        z = _get_normed_line(mat, i, x, start, end, is_sparse)
+
+        # sort chords
+        ids = None
+
+        if kwargs.get("sort", "size") == "size":
+            ids = np.argsort(z)
+        elif kwargs["sort"] == "distance":
+            remainder = 0 if num_nodes % 2 else -1
+
+            ids  = list(range(i - int(0.5*num_nodes), i))[::-1]
+            ids += list(range(i, i + int(0.5*num_nodes) + remainder + 1))
+
+            # put them back into [0, num_nodes[
+            ids = np.array(ids)
+            ids[ids < 0] += num_nodes
+            ids[ids >= num_nodes] -= num_nodes
+        else:
+            raise ValueError("Invalid `sort`: '{}'".format(kwargs["sort"]))
+
+        z0 = start
+
+        for j in ids:
+            pos[(i, j)] = (z0, z0 + z[j])
+            z0 += z[j]
+
+        start = end + pad
+
+    for i in range(len(x)):
+        start, end = arc[i]
+
+        ideogram_arc(start=start, end=end, radius=1.0, color=colors[i],
+                     width=width, alpha=alpha, ax=ax)
+
+        start, end = pos[(i, i)]
+
+        self_chord_arc(start, end, radius=1 - width - gap,
+                       chordwidth=0.7*chordwidth, color=colors[i],
+                       alpha=alpha, ax=ax)
+
+        color = colors[i]
+
+        for j in range(i):
+            cend = colors[j]
+
+            start1, end1 = pos[(i, j)]
+            start2, end2 = pos[(j, i)]
+
+            chord_arc(start1, end1, start2, end2, radius=1 - width - gap,
+                      chordwidth=chordwidth, color=colors[i], cend=cend,
+                      alpha=alpha, ax=ax, use_gradient=use_gradient)
+
+    # add names if necessary
+    if names is not None:
+        assert len(names) == num_nodes, "One name per node is required."
+
+        prop = {
+            "fontsize": kwargs.get("fontsize", 16*0.8),
+            "ha": "center",
+            "va": "center"
+        }
+
+        for pos, name in zip(nodePos, names):
+            ax.text(pos[0], pos[1], name, rotation=pos[2], **prop)
+
+    # configure axis
+    ax.set_xlim(-1.1, 1.1)
+    ax.set_ylim(-1.1, 1.1)
+
+    ax.set_aspect(1)
+    ax.axis('off')
+
+    plt.tight_layout()
+
+    if show:
+        plt.show()
+
+    return nodePos
 
 
-def polar2xy(r, theta):
-    return np.array([r*np.cos(theta), r*np.sin(theta)])
-
+# ------------ #
+# Subfunctions #
+# ------------ #
 
 def initial_path(start, end, radius, width, factor=4/3):
     ''' First 16 vertices and 15 instructions are the same for everyone '''
@@ -301,7 +491,7 @@ def chord_arc(start1, end1, start2, end2, radius=1.0, pad=2, chordwidth=0.7,
 def self_chord_arc(start, end, radius=1.0, chordwidth=0.7, ax=None,
                    color=(1,0,0), alpha=0.7):
     start, end, verts, codes = initial_path(start, end, radius, chordwidth)
-    
+
     rchord = radius * (1 - chordwidth)
 
     verts += [
@@ -323,183 +513,3 @@ def self_chord_arc(start, end, radius=1.0, chordwidth=0.7, ax=None,
         ax.add_patch(patch)
 
     return verts, codes
-
-
-def chord_diagram(mat, names=None, width=0.1, pad=2., gap=0., chordwidth=0.7,
-                  ax=None, colors=None, cmap=None, alpha=0.7,
-                  use_gradient=False, show=False, **kwargs):
-    """
-    Plot a chord diagram.
-
-    Parameters
-    ----------
-    mat : square matrix
-        Flux data, mat[i, j] is the flux from i to j
-    names : list of str, optional (default: no names)
-        Names of the nodes that will be displayed.
-    width : float, optional (default: 0.1)
-        Width/thickness of the ideogram arc.
-    pad : float, optional (default: 2)
-        Distance between two neighboring ideogram arcs. Unit: degree.
-    gap : float, optional (default: 0)
-        Distance between the arc and the beginning of the cord.
-    chordwidth : float, optional (default: 0.7)
-        Position of the control points for the chords, controlling their shape.
-    ax : matplotlib axis, optional (default: new axis)
-        Matplotlib axis where the plot should be drawn.
-    colors : list, optional (default: from `cmap`)
-        List of user defined colors or floats.
-    cmap : str or colormap object (default: viridis)
-        Colormap to use.
-    alpha : float in [0, 1], optional (default: 0.7)
-        Opacity of the chord diagram.
-    use_gradient : bool, optional (default: False)
-        Whether a gradient should be use so that chord extremities have the
-        same color as the arc they belong to.
-    **kwargs : keyword arguments
-        Available kwargs are "fontsize" and "sort" (either "size" or
-        "distance"), "zero_entry_size" (in degrees, default: 0.5).
-    """
-    import matplotlib.pyplot as plt
-
-    if ax is None:
-        _, ax = plt.subplots()
-
-    # mat[i, j]:  i -> j
-    num_nodes = len(mat)
-
-    # copy matrix and set a minimal value for visibility of zero fluxes
-    mat = mat.copy()
-
-    min_deg  = kwargs.get("zero_entry_size", 0.5)
-    min_deg *= mat.sum() / (360 - num_nodes*pad)
-
-    zeros = np.argwhere(mat == 0)
-
-    for (i, j) in zeros:
-        if mat[j, i] != 0:
-            mat[i, j] = min_deg
-
-    # sum over rows
-    x = mat.sum(axis=1)
-    
-    # configure colors
-    if colors is None:
-        colors = np.linspace(0, 1, num_nodes)
-
-    if cmap is None:
-        cmap = "viridis"
-
-    if isinstance(colors, (Sequence, np.ndarray)):
-        assert len(colors) == num_nodes, "One color per node is required."
-
-        # check color type
-        first_color = colors[0]
-
-        if isinstance(first_color, (int, float, np.integer)):
-            cm = plt.get_cmap(cmap)
-            colors = cm(colors)[:, :3]
-        else:
-            colors = [ColorConverter.to_rgb(c) for c in colors]
-    else:
-        raise ValueError("`colors` should be a list.")
-
-    # find position for each start and end
-    y = x / np.sum(x).astype(float) * (360 - pad*len(x))
-
-    pos = {}
-    arc = []
-    nodePos = []
-    start = 0
-
-    for i in range(len(x)):
-        end = start + y[i]
-        arc.append((start, end))
-        angle = 0.5*(start+end)
-
-        if -30 <= angle <= 210:
-            angle -= 90
-        else:
-            angle -= 270
-
-        nodePos.append(
-            tuple(polar2xy(1.1, 0.5*(start + end)*np.pi/180.)) + (angle,))
-
-        z = (mat[i, :] / x[i].astype(float)) * (end - start)
-
-        # sort chords
-        ids = None
-
-        if kwargs.get("sort", "size") == "size":
-            ids = np.argsort(z)
-        elif kwargs["sort"] == "distance":
-            remainder = 0 if num_nodes % 2 else -1
-
-            ids  = list(range(i - int(0.5*num_nodes), i))[::-1]
-            ids += list(range(i, i + int(0.5*num_nodes) + remainder + 1))
-
-            # put them back into [0, num_nodes[
-            ids = np.array(ids)
-            ids[ids < 0] += num_nodes
-            ids[ids >= num_nodes] -= num_nodes
-        else:
-            raise ValueError("Invalid `sort`: '{}'".format(kwargs["sort"]))
-
-        z0 = start
-
-        for j in ids:
-            pos[(i, j)] = (z0, z0 + z[j])
-            z0 += z[j]
-
-        start = end + pad
-
-    for i in range(len(x)):
-        start, end = arc[i]
-
-        ideogram_arc(start=start, end=end, radius=1.0, color=colors[i],
-                     width=width, alpha=alpha, ax=ax)
-
-        start, end = pos[(i, i)]
-
-        self_chord_arc(start, end, radius=1 - width - gap,
-                       chordwidth=0.7*chordwidth, color=colors[i],
-                       alpha=alpha, ax=ax)
-
-        color = colors[i]
-
-        for j in range(i):
-            cend = colors[j]
-
-            start1, end1 = pos[(i, j)]
-            start2, end2 = pos[(j, i)]
-
-            chord_arc(start1, end1, start2, end2, radius=1 - width - gap,
-                      chordwidth=chordwidth, color=colors[i], cend=cend,
-                      alpha=alpha, ax=ax, use_gradient=use_gradient)
-
-    # add names if necessary
-    if names is not None:
-        assert len(names) == num_nodes, "One name per node is required."
-
-        prop = {
-            "fontsize": kwargs.get("fontsize", 16*0.8),
-            "ha": "center",
-            "va": "center"
-        }
-
-        for pos, name in zip(nodePos, names):
-            ax.text(pos[0], pos[1], name, rotation=pos[2], **prop)
-
-    # configure axis
-    ax.set_xlim(-1.1, 1.1)
-    ax.set_ylim(-1.1, 1.1)
-
-    ax.set_aspect(1)
-    ax.axis('off')
-
-    plt.tight_layout()
-
-    if show:
-        plt.show()
-
-    return nodePos
